Move ParseVolumes to its own file
It was in the ports file.
